backend/requirements.txt
Django==5.2.9
djangorestframework
django-cors-headers
djangorestframework-gis
djangorestframework-simplejwt
psycopg2-binary
django-filter
drf-spectacular